+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1] - 2026-02-02
+
+   - ANIMATION PUBLISHER:
+      - Improve logging
+      - Check character rig version is the last one. If not, notify artist to change it.
+      - Create a publish on SG to track
+
+## [1.1.0] - 2026-02-02
+
+   - ANIMATION PUBLISHER
+
 ## [1.0.18] - 2026-01-26
 
    - SPLITTER: